+++ src/monorepo-npm-publisher/CHANGELOG.md
@@ -9,4 +9,4 @@
 
 # 1.1.0
 
-- Internal dependencies upgraded and `@kiwicom/babel-preset-kiwicom` bumped to version 3.0.0 which now supports dynamic `import()` syntax.
+- Internal dependencies upgraded and `@adeira/babel-preset-adeira` bumped to version 3.0.0 which now supports dynamic `import()` syntax.